[mmptest] Port the link-xml-serialization test to an NUnit LinkAll test. Partial fix for #4975.
Fixes part of https://github.com/xamarin/xamarin-macios/issues/4975.
