--- tests/mmptest/regression/link-xml-serialization/Main.cs
@@ -1,18 +1,0 @@
-using System;
-using System.Drawing;
-using MonoMac.Foundation;
-using MonoMac.AppKit;
-using MonoMac.ObjCRuntime;
-
-namespace linkxmlserialization
-{
-	class MainClass
-	{
-		static void Main (string [] args)
-		{
-			NSApplication.Init ();
-			NSApplication.Main (args);
-		}
-	}
-}	
-
